+++ ocaml/libs/backtrace/test/log.mli
@@ -1,0 +1,4 @@
+val with_thread_associated : string -> ('a -> 'b) -> 'a -> 'b
+
+val with_thread_associated_old : string -> ('a -> 'b) -> 'a -> 'b
+(** Uses V1.with_backtrace *)