+++ ocaml/xe-cli/newcli.ml
@@ -75,7 +75,7 @@ exception Usage
 let usage () =
   let help =
     Printf.sprintf
-      {|Usage: 
+      {|Usage:
   %s <command>
     [ -s <server> ]            XenServer host
     [ -p <port> ]              XenServer port number
@@ -84,7 +84,9 @@ let usage () =
     [ --nossl ]                Disable SSL/TLS
     [ --debug ]                Enable debug output
     [ --debug-on-fail ]        Enable debug output only on failure
+    [ --minimal ]              Output only values in a comma-separated lists
     [ --traceparent <value> ]  Distributed tracing context
+    [ --trace ]                Show task stacktrace on failure
     [ <other arguments> ... ]  Command-specific options
 
 A full list of commands can be obtained by running